+++ packages/breadboard-ui/src/elements/editor/editor.ts
@@ -54,13 +54,14 @@ export class Editor extends LitElement {
   kits: Kit[] = [];
 
   @property()
-  nodeCount = 0;
-
-  @property()
-  edgeCount = 0;
-
-  @property()
   editable = false;
+
+  /**
+   * Used to attempt a graph re-render. This isn't guaranteed to happen, though,
+   * as it depends on whether the editor was expecting the change.
+   */
+  @property()
+  renderCount = 0;
 
   @property()
   highlightedNodeId: string | null = null;
@@ -84,6 +85,7 @@ export class Editor extends LitElement {
   #onGraphNodeDeleteBound = this.#onGraphNodeDelete.bind(this);
   #top = 0;
   #left = 0;
+  #expectingRefresh = false;
 
   static styles = css`
     :host {
@@ -379,7 +381,7 @@ export class Editor extends LitElement {
     super.disconnectedCallback();
   }
 
-  protected willUpdate(
+  protected updated(
     changedProperties:
       | PropertyValueMap<{
           loadInfo: LoadArgs;
@@ -389,9 +391,10 @@ export class Editor extends LitElement {
       | Map<PropertyKey, unknown>
   ): void {
     const shouldProcessGraph =
-      changedProperties.has("loadInfo") ||
-      changedProperties.has("nodeCount") ||
-      changedProperties.has("edgeCount");
+      changedProperties.has("loadInfo") || this.#expectingRefresh;
+
+    this.#expectingRefresh = false;
+
     if (shouldProcessGraph && this.loadInfo && this.loadInfo.graphDescriptor) {
       this.#processGraph(this.loadInfo.graphDescriptor);
     }
@@ -407,7 +410,7 @@ export class Editor extends LitElement {
 
   #onGraphEdgeAttach(evt: Event) {
     const { edge } = evt as GraphNodeEdgeAttach;
-    this.edgeCount = -1;
+    this.#expectingRefresh = true;
     this.dispatchEvent(
       new EdgeChangeEvent("add", {
         from: edge.from.descriptor.id,
@@ -420,7 +423,7 @@ export class Editor extends LitElement {
 
   #onGraphEdgeDetach(evt: Event) {
     const { edge } = evt as GraphNodeEdgeDetach;
-    this.edgeCount = -1;
+    this.#expectingRefresh = true;
     this.dispatchEvent(
       new EdgeChangeEvent("remove", {
         from: edge.from.descriptor.id,
@@ -433,7 +436,7 @@ export class Editor extends LitElement {
 
   #onGraphEdgeChange(evt: Event) {
     const { fromEdge, toEdge } = evt as GraphNodeEdgeChange;
-    this.edgeCount = -1;
+    this.#expectingRefresh = true;
     this.dispatchEvent(
       new EdgeChangeEvent("remove", {
         from: fromEdge.from.descriptor.id,
@@ -455,8 +458,7 @@ export class Editor extends LitElement {
 
   #onGraphNodeDelete(evt: Event) {
     const { id } = evt as GraphNodeDelete;
-    this.edgeCount = -1;
-    this.nodeCount = -1;
+    this.#expectingRefresh = true;
     this.dispatchEvent(new NodeDeleteEvent(id));
   }
 
@@ -482,6 +484,7 @@ export class Editor extends LitElement {
     // Store the middle of the node for later.
     this.#graph.setNodeLayoutPosition(id, { x, y });
 
+    this.#expectingRefresh = true;
     this.dispatchEvent(new NodeCreateEvent(id, data));
 
     this.nodeValueBeingEdited = {
@@ -576,16 +579,7 @@ export class Editor extends LitElement {
 
     const configuration = node.configuration() || {};
     const details = (async () => {
-      const { inputSchema } = await node.describe();
-
-      if (!inputSchema.properties) {
-        return html`<div id="properties">
-          <div id="node-properties">
-            Unable to configure node - no schema properties provided.
-          </div>
-        </div>`;
-      }
-
+      const { inputs } = await node.ports();
       return html` <div id="properties">
         <div id="node-properties">
           <form @submit=${this.#onFormSubmit}>
@@ -625,44 +619,52 @@ export class Editor extends LitElement {
               />
               <fieldset>
                 <legend>Configuration</legend>
-                ${Object.keys(inputSchema.properties).length === 0
-                  ? html`No configurable properties`
-                  : nothing}
-                ${map(
-                  Object.entries(inputSchema.properties),
-                  ([name, schema]) => {
-                    let input;
-                    switch (schema.type) {
-                      case "object": {
-                        input = `Object types are not supported yet.`;
-                        break;
-                      }
+                ${map(inputs.ports, (port) => {
+                  if (port.star) return;
+                  const schema = port.schema || {};
+                  const name = port.name;
+                  const value =
+                    configuration[name] ??
+                    schema.examples ??
+                    schema.default ??
+                    "";
 
-                      // TODO: Fill out more types.
-                      default: {
-                        const value =
-                          configuration[name] ??
-                          schema.examples ??
-                          schema.default ??
-                          "";
+                  let input;
+                  const type = port.schema?.type || "string";
+                  switch (type) {
+                    case "object": {
+                      // TODO: Implement object editor.
+                      // Use "format" to distinguish what type of "object",
+                      // e.g. "format": "schema" for a JSON schema.
+                      input = html`<div>
+                        <textarea
+                          style="height: 200px; width: 100%"
+                          name="schema"
+                        >
+${JSON.stringify(value, null, 2)}</textarea
+                        >
+                      </div>`;
+                      break;
+                    }
 
-                        // prettier-ignore
-                        input = html`<div
+                    // TODO: Fill out more types.
+                    default: {
+                      // prettier-ignore
+                      input = html`<div
                             contenteditable="plaintext-only"
                             data-id="${name}"
                           >${value}</div>`;
-                        break;
-                      }
+                      break;
                     }
-
-                    return html`<div class="configuration-item">
-                      <label title="${schema.description}" for="${name}"
-                        >${name}:
-                      </label>
-                      ${input}
-                    </div>`;
                   }
-                )}
+
+                  return html`<div class="configuration-item">
+                    <label title="${schema.description}" for="${name}"
+                      >${name}:
+                    </label>
+                    ${input}
+                  </div>`;
+                })}
               </fieldset>
             </div>
           </form>
@@ -727,9 +729,17 @@ export class Editor extends LitElement {
         continue;
       }
 
+      if (name === "schema") {
+        configuration[name] = JSON.parse(value);
+        continue;
+      }
+
       configuration[name] = value;
     }
 
+    console.log("New configuration", configuration);
+
+    this.#expectingRefresh = true;
     this.dispatchEvent(new NodeUpdateEvent(id, configuration));
 
     // Close out the panel via removing the active node marker.